Initial Plugin architecture. Read more below plz
This is an initial commit for the plugin architecture. It is not complete yet,
and shouldn't be used. The goal is to have loadable plugins for each major
sub-system where we might want multiple implementations. For example, in the
future we might have different Rendering systems (Vulkan-based, 2D, 1D, DirecX
etc.), or even use external libraries/APIs for sound system, animation, physics
etc. The engine should expect having at least one working interface for each
system, and that could be loadable on-the-fly with this plugin system. There's
far more work to be done, more instructions in the @todo comments.